+++ NEWS.md
@@ -1,3 +1,6 @@
+# TreeDist 0.0.0.9004 (development)
+ - Add TreeDistPlot function
+
 # TreeDist 0.0.0.9000 (development)
 
  - Extracted package from `TreeSearch` 0.3.2.9005.